refactor: move mockCampaigns data to mock/campaignData.ts and import in CampaignSection
