+++ mock/campaignData.ts
@@ -1,76 +1,110 @@
-import { MockCampaign } from "@/types/mock";
+export interface Campaign {
+  id: string;
+  title: string;
+  description: string;
+  ngoName: string;
+  location: string;
+  category: string;
+  targetAmount: number;
+  raisedAmount: number;
+  donorCount: number;
+  daysLeft: number;
+  image: string;
+  urgent: boolean;
+  featured: boolean;
+}
 
-export const mockCampaigns: MockCampaign[] = [
+const mockCampaigns: Campaign[] = [
   {
     id: "1",
-    title: "Help Build Clean Water Wells in Rural Communities",
-    description:
-      "We are raising funds to construct clean water wells in remote villages that lack access to safe drinking water. Your donation will directly impact hundreds of families by providing them with a reliable source of clean water, improving health outcomes and reducing water-borne diseases.",
-    type: "charity",
-    donationPercent: 68,
-    raisedAmount: 34000,
-    goalAmount: 50000,
-    image:
-      "https://images.unsplash.com/photo-1594736797933-d0300ce27667?w=500&h=300&fit=crop",
+    title: "Emergency Relief for Flood Victims",
+    description: "Providing immediate shelter, food, and medical aid to families affected by recent flooding in rural communities.",
+    ngoName: "Disaster Relief Foundation",
+    location: "Bangladesh",
+    category: "Emergency Relief",
+    targetAmount: 50000,
+    raisedAmount: 32500,
+    donorCount: 245,
+    daysLeft: 12,
+    image: "flood.png",
+    urgent: true,
+    featured: true
   },
   {
     id: "2",
-    title: "Emergency Surgery Fund for 8-Year-Old Emma",
-    description:
-      "Little Emma needs urgent heart surgery that her family cannot afford. The procedure is critical for her survival and future quality of life. Every dollar donated brings us closer to giving Emma the chance at a healthy, happy childhood she deserves.",
-    type: "medical",
-    donationPercent: 42,
-    raisedAmount: 21000,
-    goalAmount: 50000,
-    image:
-      "https://images.unsplash.com/photo-1559757148-5c350d0d3c56?w=500&h=300&fit=crop",
+    title: "Clean Water Wells for Remote Villages",
+    description: "Building sustainable water infrastructure to provide clean drinking water to 5 remote villages.",
+    ngoName: "Water for Life",
+    location: "Kenya",
+    category: "Water & Sanitation",
+    targetAmount: 75000,
+    raisedAmount: 45000,
+    donorCount: 189,
+    daysLeft: 28,
+    image: "flood.png",
+    urgent: false,
+    featured: true
   },
   {
     id: "3",
-    title: "Scholarship Fund for Underprivileged Students",
-    description:
-      "Help us provide educational opportunities for talented students from low-income families. This scholarship fund will cover tuition, books, and living expenses for students pursuing higher education, breaking the cycle of poverty through education.",
-    type: "education",
-    donationPercent: 85,
-    raisedAmount: 17000,
-    goalAmount: 20000,
-    image:
-      "https://images.unsplash.com/photo-1523580494863-6f3031224c94?w=500&h=300&fit=crop",
+    title: "Education Scholarships for Girls",
+    description: "Supporting 100 girls from low-income families to complete their secondary education with full scholarships.",
+    ngoName: "Girls Education Initiative",
+    location: "India",
+    category: "Education",
+    targetAmount: 30000,
+    raisedAmount: 18750,
+    donorCount: 156,
+    daysLeft: 45,
+    image: "flood.png",
+    urgent: false,
+    featured: false
   },
   {
     id: "4",
-    title: "Wildlife Rescue and Rehabilitation Center",
-    description:
-      "Support our mission to rescue, rehabilitate, and release injured wildlife back into their natural habitats. Your donations help us provide medical care, food, and shelter for animals in need while supporting conservation efforts in our local ecosystem.",
-    type: "animal",
-    donationPercent: 73,
-    raisedAmount: 29200,
-    goalAmount: 40000,
-    image:
-      "https://images.unsplash.com/photo-1547036967-23d11aacaee0?w=500&h=300&fit=crop",
+    title: "Mobile Medical Clinic for Rural Areas",
+    description: "Funding a mobile medical unit to provide healthcare services to underserved rural communities.",
+    ngoName: "Healthcare Access Network",
+    location: "Guatemala",
+    category: "Healthcare",
+    targetAmount: 85000,
+    raisedAmount: 51000,
+    donorCount: 298,
+    daysLeft: 21,
+    image: "flood.png",
+    urgent: false,
+    featured: true
   },
   {
     id: "5",
-    title: "Hurricane Relief Fund - Immediate Aid Needed",
-    description:
-      "Communities devastated by Hurricane Maria need immediate assistance. We are providing emergency supplies, temporary shelter, and rebuilding support to affected families. Time is critical - help us deliver aid to those who need it most.",
-    type: "emergency",
-    donationPercent: 91,
-    raisedAmount: 45500,
-    goalAmount: 50000,
-    image:
-      "https://images.unsplash.com/photo-1547036967-23d11aacaee0?w=500&h=300&fit=crop",
+    title: "Reforestation Project - Plant 10,000 Trees",
+    description: "Combat climate change by planting native trees and restoring degraded forest ecosystems.",
+    ngoName: "Green Earth Foundation",
+    location: "Brazil",
+    category: "Environment",
+    targetAmount: 25000,
+    raisedAmount: 19500,
+    donorCount: 412,
+    daysLeft: 35,
+    image: "flood.png",
+    urgent: false,
+    featured: false
   },
   {
     id: "6",
-    title: "Community Garden Initiative",
-    description:
-      "Transform vacant lots into thriving community gardens that provide fresh produce to local families while building stronger neighborhood connections. This initiative will create sustainable food sources and educational opportunities for children and adults alike.",
-    type: "community",
-    donationPercent: 56,
-    raisedAmount: 11200,
-    goalAmount: 20000,
-    image:
-      "https://images.unsplash.com/photo-1416879595882-3373a0480b5b?w=500&h=300&fit=crop",
-  },
+    title: "Winter Clothing Drive for Homeless",
+    description: "Providing warm clothing, blankets, and winter essentials to homeless individuals during harsh winter months.",
+    ngoName: "Shelter & Care Foundation",
+    location: "Canada",
+    category: "Social Welfare",
+    targetAmount: 15000,
+    raisedAmount: 8900,
+    donorCount: 87,
+    daysLeft: 18,
+    image: "flood.png",
+    urgent: true,
+    featured: false
+  }
 ];
+
+export default mockCampaigns;
